Rename secrets to config.
