+++ lockbox/.gitignore
@@ -1,1 +1,1 @@
-secrets.py
+config.py
